api/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
slowapi>=0.1.9
pydantic>=2.7.0
python-multipart>=0.0.9
pdfplumber>=0.11.0
python-docx>=1.1.2
anthropic>=0.31.0
openai>=1.35.0
google-generativeai>=0.7.2
zhipuai>=2.1.1.20240902
absl-py>=1.0.0
aiohttp>=3.8.0
async_timeout>=4.0.0
exceptiongroup>=1.1.0
google-genai>=1.39.0
google-cloud-storage>=2.14.0
ml-collections>=0.1.0
more-itertools>=8.0.0
numpy>=1.20.0
pandas>=1.3.0
python-dotenv>=0.19.0
PyYAML>=6.0
regex>=2023.0.0
requests>=2.25.0
tqdm>=4.64.0
typing-extensions>=4.0.0
